+++ entry.jsx
@@ -1,8 +1,3 @@
-import { UserInterface } from "./lib/Views/UserInterface";
-import Variables from "./lib/Styles/variables.scss";
-import { createRoot } from "react-dom/client";
+import { renderUi } from "./lib/Views/root";
 
-const container = document.getElementById("ui");
-const root = createRoot(container);
-
-root.render(<UserInterface themeOverrides={Variables} />);
+renderUi();
